- move get config function to a common place to be used by conn.c and notifications.c
